Reference

PSF Models

PointSpreadFunctions.AbstractPSF — Type

Instances of sub-types of PointSpreadFunctions.AbstractPSF{N} are used to store the parameters of various Point Spread Functions (PSF) models, N is the number of parameters. Let P be such an instance, then:

P([T,] r)
P([T,] x, y)

yield an approximation of the PSF at a distance r or at a position (x,y) relative to the position of the point source.

Use index notation to retrieve the parameters as an N-tuple or a specific parameters:

P[:]  # yields all parameters
P[i]  # yields i-th parameter

The PSF is normalized such that its peak value (usually at the center) is equal to one. The rationale is that normalization by the peak value does not depend on the dimensionality while the integral of the PSF does depend on the dimensionality.

See also AiryPSF, CauchyPSF, GaussianPSF, MoffatPSF.

PointSpreadFunctions.AiryPSF — Type

P = AiryPSF(lim [, eps=0])

defines the Point Spread Function (PSF) P of a circular pupil. Argument lim is the distance in the focal plane corresponding to the diffraction limit. Assuming the distance is given in angular units (radians), lim = λ/D with λ the wavelength and D the pupil diameter (both expressed in the same units). Optional argument eps is the ratio of the size of the central obscuration to that of the pupil (necessarily 0 ≤ eps < 1). If not specified, eps = 0 is assumed (unobstructed pupil). Then:

P([T,] r)
P([T,] x, y)

yield the PSF at distance r or position (x,y) relative to the position of the source and normalized so that the peak value of the PSF is one. Optional argument T is the floating-point type of the result. Arguments lim, r, x and y are in the same units.

See also CauchyPSF, GaussianPSF, MoffatPSF.

PointSpreadFunctions.CauchyPSF — Type

P = CauchyPSF(fwhm)
P([T,] r)
P([T,] x, y)

defines a Cauchy Point Spread Function (PSF) P (or Lorentzian) of full width at half maximum fwhm which can be used to compute the PSF at distance r or position (x,y) relative to the position of the source and normalized so that the peak value of the PSF is one. Optional argument T is the floating-point type of the result. Arguments fwhm, r, x and y are in the same units.

See also AiryPSF, GaussianPSF, MoffatPSF.

PointSpreadFunctions.GaussianPSF — Type

P = GaussianPSF(fwhm)
P([T,] r)
P([T,] x, y)

defines a Gaussian Point Spread Function (PSF) P of full width at half maximum fwhm which can be used to compute the PSF at distance r or position (x,y) relative to the position of the source and normalized so that the peak value of the PSF is one. Optional argument T is the floating-point type of the result. Arguments fwhm, r, x and y are in the same units.

See also AiryPSF, CauchyPSF, MoffatPSF.

PointSpreadFunctions.MoffatPSF — Type

P = MoffatPSF(fwhm, beta)
P([T,] r)
P([T,] x, y)

defines a Moffat PSF P of full width at half maximum fwhm and exponent beta which can be used to compute the PSF at distance r or position (x,y) relative to the position of the source and normalized so that the peak value of the PSF is one. Optional argument T is the floating-point type of the result. Arguments fwhm, r, x and y are in the same units.

See also AiryPSF, CauchyPSF, GaussianPSF.

Methods

PointSpreadFunctions.getfwhm — Function

getfwhm(P)

yields the full width at half maximum (FWHM) of point spread function P. Computing the result may require some calculations, so it is better to save the result if this value must be used several times.

PointSpreadFunctions.findzeros — Function

findzeros([T=Float64,] P, n; kwds...)

yield the n-th zero of Point Spread Function (PSF) P. If n is a vector of integers, a vector with the corresponding zeros is returned. For example:

findzeros(P, 1:3)

yields the first 3 zeros of P.

Optional argument T can be used to specify the floating-point type of the result. The keywords of OptimPack.fzero can be specified.

PointSpreadFunctions.fit — Function

fit(psf, pos, [wgt,] dat; nonnegative=false) -> psf′, pos′

fits a given Point Spread Function (PSF) model on data dat with (optional) respective weights wgt. Argument psf is a PSF instance to specify which PSF model to use and its initial parameters. Argument pos is the initial PSF position. The result is a 2-tuple with the fitted PSF model and position.

The fit is carried out by OptimPack.Powell.Newuoa.minimize method. The initial parameters should be close enough to the solution for the fit to behave correctly.

Keyword nonnegative indicates whether the intensity of the PSF should be nonnegative or not. Keyword rho = (rhobeg,rhoend) may be used to specify the initial and final precision on the parameters. Additional keywords may be specified and are passed to the minimizer.