List of Functions

Parameters

see Workflow

PSF generation

see Workflow

Pupils

Missing docstring.

Missing docstring for PSFs.pupil_θ(sz, pp::PSFParams, sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.pupil_ϕ(sz, pp::PSFParams, sampling). Check Documenter's build log for details.

PointSpreadFunctions.aplanatic_factor — Method

aplanatic_factor(sz, pp::PSFParams, sampling)

returns the aplanatic factor as specified in pp.aplanatic as a pupil array.

PointSpreadFunctions.pupil_xyz — Function

pupil_xyz(sz, pp, sampling=nothing)

creates a pupil with electric field distributions in XYZ. Returns a 4D dataset with the electric field components along the 4th dimension. #Arguments

'sz': size of the pupil in pixels

'pp': the PSFParam structure with all the PSF parameters

'sampling': the pixel sampling in the same units as the wavelength

'is_proj': defines whether the pupil is to be interpreted as the projection of the McCutchen pupil or not. This yields a by 1 ./cos(Theta) modified aplanatic factor.

Missing docstring.

Missing docstring for PSFs.field_xyz(sz, pp, sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.field_xy_to_xyz(field,pp,sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.field_pupil. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.get_propagator(sz,pp,sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.get_propagator_gradient(prop_phase, scalar, xy_scale). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.apply_propagators(pupil, z_planes, pp::PSFParams; sampling=nothing). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.get_zernike_pupil_phase(sz, pp, sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.get_zernike_pupil(sz, pp, sampling). Check Documenter's build log for details.

Polarization

These functions can be conveniently supplied to PSFParams() via the named argument polarization

PointSpreadFunctions.pol_scalar — Function

pol_scalar(T, xypos)

ignores polarization aspects in the calculation but calculates only based on (high-NA) scalar theory. This is a lot faster but not as accurate.

PointSpreadFunctions.pol_scalar_spiral — Function

pol_scalar_spiral(T, xypos)

ignores polarization aspects in the calculation but calculates only based on (high-NA) scalar theory. This version still includes a (scalar) phase spiral. This is a lot faster but not as accurate.

PointSpreadFunctions.pol_x — Function

pol_x(T, xypos)

assumes x-polarization in illumination or an x-oriented polarizer in detection. In a high-NA objective this is converted into XYZ electric fields at the focus.

PointSpreadFunctions.pol_y — Function

pol_y(T, xypos)

assumes y-polarization in illumination or an x-oriented polarizer in detection. In a high-NA objective this is converted into XYZ electric fields at the focus.

PointSpreadFunctions.pol_circ — Function

pol_circ(T, xypos)

assumes circular polarization in illumination or an x-oriented polarizer in detection. In a high-NA objective this is converted into XYZ electric fields at the focus.

PointSpreadFunctions.pol_circ_spiral — Function

pol_circ_spiral(T, xypos)

assumes circular polarization in illumination or an x-oriented polarizer in detection. This version includes phase spiral defining the local (xypos-dependent) phase of both x and y polarization.

Aplanatic factors

PointSpreadFunctions.aplanatic_detection — Function

aplanatic_detection = (θ) -> sqrt.(max.(0,cos.(θ)))

This is the aplanatic factor typically used in detection of fluorescence of (randomly oriented) fluorophores.

PointSpreadFunctions.aplanatic_illumination — Function

aplanatic_illumination = (θ) -> sqrt.(max.(0,cos.(θ)))

This is the aplanatic factor typically used in illumination of (randomly oriented) fluorophores. Note that it is identical to detection.

PointSpreadFunctions.aplanatic_const — Function

aplanatic_const = (θ) -> one.(eltype(θ))

This is a constant aplanatic factor

PointSpreadFunctions.aplanatic_illumination_flux — Function

aplanatic_illumination_flux = (θ) -> max.(0,cos.(θ))

This refers the aplanatic factor if interested in the flux through a detector perpendicular to the optical axis.

Utilities

Missing docstring.

Missing docstring for PSFs.amp_to_int(field). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.has_z_symmetry(pp::PSFParams). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.get_Abbe_limit(pp::PSFParams). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.get_required_amp_sampling(sz::NTuple, pp::PSFParams). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.get_Ewald_sampling(sz::NTuple, pp::PSFParams). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.get_McCutchen_kz_center(sz, pp::PSFParams, sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.limit_kz(ft_shell, pp::PSFParams, sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.sinc_r(sz::NTuple, pp::PSFParams; sampling=nothing). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.jinc_r_2d(sz::NTuple, pp::PSFParams; sampling=nothing). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.my_disc(sz, pp). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.iftz(arr). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.theta_z(sz). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.k_0(pp::PSFParams). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.k_pupil(pp::PSFParams). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.k_dz(pp::PSFParams). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.k_scale(sz, pp::PSFParams, sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.k_pupil_pos(sz, pp::PSFParams, sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.k_0_pos(sz, pp::PSFParams, sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.k_r(sz, pp::PSFParams, sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.k_xy(sz,pp,sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.k_xy_rel_pupil(sz,pp,sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.check_amp_sampling_xy(sz, pp,sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.check_amp_sampling_z(sz, pp,sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.check_amp_sampling(sz, pp,sampling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSFs.check_amp_sampling_sincr(sz, pp,sampling). Check Documenter's build log for details.