Reference

PSF Models

PointSpreadFunctions.AbstractPSF — Type

Instances of sub-types of PointSpreadFunctions.AbstractPSF{N} are used to store the parameters of various Point Spread Functions (PSF) models, N is the number of parameters. Let P be such an instance, then:

P([T,] r)
P([T,] x, y)

yield an approximation of the PSF at a distance r or at a position (x,y) relative to the position of the point source.

Use index notation to retrieve the parameters as an N-tuple or a specific parameters:

P[:]  # yields all parameters
P[i]  # yields i-th parameter

The PSF is normalized such that its peak value (usually at the center) is equal to one. The rationale is that normalization by the peak value does not depend on the dimensionality while the integral of the PSF does depend on the dimensionality.

See also AiryPSF, CauchyPSF, GaussianPSF, MoffatPSF.

PointSpreadFunctions.AiryPSF — Type

P = AiryPSF(lim [, ρ=0])

defines the Point Spread Function (PSF) P of a circular pupil. Argument lim is the distance in the focal plane corresponding to the diffraction limit. Assuming the distance is given in angular units (radians), lim = λ/D with λ the wavelength and D the pupil diameter (both expressed in the same units). Optional argument ρ is the ratio of the size of the central obscuration to that of the pupil (necessarily 0 ≤ ρ < 1). If not specified, ρ = 0 is assumed (unobstructed pupil). Then:

P([T,] r)
P([T,] x, y)

yield the PSF at distance r or position (x,y) relative to the position of the source and normalized so that the peak value of the PSF is one. Optional argument T is the floating-point type of the result. Arguments lim, r, x and y are in the same units.

The peak-normalized Airy PSF at distance r is given by:

P(r) = ((a/(c*r))*(J1(c*r) - ρ*J1(ρ*c*r)))^2

with a = 2/(1 - ρ^2), c = π/lim and J1 the Bessel function of the first kind.

See also CauchyPSF, GaussianPSF, MoffatPSF.

PointSpreadFunctions.CauchyPSF — Type

P = CauchyPSF(fwhm)

defines a Cauchy Point Spread Function (PSF) P (or Lorentzian) of full width at half maximum fwhm which can be used as:

P([T,] r)
P([T,] x, y)

to compute the PSF at distance r or position (x,y) relative to the position of the source and normalized so that the peak value of the PSF is one. Optional argument T is the floating-point type of the result. Arguments fwhm, r, x and y are in the same units.

The peak-normalized Cauchy PSF at distance r is given by:

P(r) = q/(q + r^2)

with q = (fwhm/2)^2.

See also AiryPSF, GaussianPSF, MoffatPSF.

PointSpreadFunctions.GaussianPSF — Type

P = GaussianPSF(fwhm)

defines a Gaussian Point Spread Function (PSF) P of full width at half maximum fwhm which can be used as:

P([T,] r)
P([T,] x, y)

to compute the PSF at distance r or position (x,y) relative to the position of the source and normalized so that the peak value of the PSF is one. Optional argument T is the floating-point type of the result. Arguments fwhm, r, x and y are in the same units.

The peak-normalized Gaussian PSF at distance r is given by:

P(r) = exp(q*r^2)

with q = -4*log(2)/fwhm^2.

See also AiryPSF, CauchyPSF, MoffatPSF.

PointSpreadFunctions.MoffatPSF — Type

P = MoffatPSF(fwhm, beta)

defines a Moffat PSF P of full width at half maximum fwhm and exponent beta which can be used as:

P([T,] r)
P([T,] x, y)

to compute the PSF at distance r or position (x,y) relative to the position of the source and normalized so that the peak value of the PSF is one. Optional argument T is the floating-point type of the result. Arguments fwhm, r, x and y are in the same units.

The peak-normalized Moffat PSF at distance r is given by:

P(r) = (1 + q*r^2)^(-beta)

with q = (2^(1/beta) - 1)*(2/fwhm)^2.

See also AiryPSF, CauchyPSF, GaussianPSF.

Methods

PointSpreadFunctions.getfwhm — Function

getfwhm(P)

yields the full width at half maximum (FWHM) of point spread function P. Computing the result may require some calculations, so it is better to save the result if this value must be used several times.

PointSpreadFunctions.findzeros — Function

findzeros([T=Float64,] P, n; kwds...)

yield the n-th zero of Point Spread Function (PSF) P. If n is a vector of integers, a vector with the corresponding zeros is returned. For example:

findzeros(P, 1:3)

yields the first 3 zeros of P.

Optional argument T can be used to specify the floating-point type of the result. The keywords of OptimPack.fzero can be specified.

PointSpreadFunctions.fit — Function

fit(psf, pos, [wgt,] dat[, roi]; nonnegative=false) -> psf′, pos′

fits a given Point Spread Function (PSF) model on data dat with (optional) respective weights wgt. Argument psf is a PSF instance to specify which PSF model to use and its initial parameters. Argument pos is the initial PSF position. The result is a 2-tuple with the fitted PSF model and position.

The fit is carried out by OptimPack.Powell.Newuoa.minimize method. The initial parameters should be close enough to the solution for the fit to behave correctly.

Keyword nonnegative indicates whether the intensity of the PSF should be nonnegative or not. Keyword rho = (rhobeg,rhoend) may be used to specify the initial and final precision on the parameters. Additional keywords may be specified and are passed to the minimizer.

An optional Region Of Interest (ROI) may be specified by argument roi as an N-tuple of index ranges or colons. This is useful to limit the region used to perform the fit. For example:

fit(GaussianPSF(2.5), (23,12), dat, (10:30, :))

will fit a Gaussian PSF model on sub-array dat[10:30,:]. The advantages of specifying a ROI is that the relative position of the ROI is taken into account in the initial and final position and that a view is used instead of extracting a sub-array. Specifying a ROI also works for weighted data. Non-rectangular ROIs can be emulated by having weights equal to zero where data should be ignored.

To simply localize the PSF without fitting its parameters, wrap the PSF model with the immutable method:

fit(immutable(psf), pos, [wgt,] dat[, roi]; nonnegative=false)

Note

The fit and immutable methods are not exported (yet). They must be prefixed or explicitely exported. For example:

using PointSpreadFunctions
using PointSpreadFunctions: fit, immutable

Base.isvalid — Method

isvalid(psf) -> bool

yields whether the parameters of the point spread function psf are correct. The check_structure method is similar but throws for invalid parameters.

Missing docstring.

Missing docstring for check_structure(::PointSpreadFunctions.AbstractPSF). Check Documenter's build log for details.