PointSpreadFunction parameters

PointSpreadFunctions.PSFParams — Type

PSFParams

This structure stores all the general parameters needed to calculate the point spread function. Only pixel-pitch and image size information is handles seperately.

Structure Members:

λ: Vacuum wavelength

NA: numerical aperture

n: refractive index of the embedding AND immersion medium

dtype: real-valued data type to generate PSF for

mode: microscopy mode to calculate PSF for. See the constructor PSFParams() for more details.

polarization: a function calculating the polarization from a given Tuple of relative-k pupil vectors. The constructur argument is pol.

aplanatic: aplanatic factor. Provided as a function of angle θ

method: the method of calculation

FFTPlan: information on how to calculate the FFTW plan

See also:

PSFParams()

Example:

julia> using FFTW, PointSpreadFunctions

julia> ppm = PSFParams(0.580, 1.4, 1.518;pol=pol_scalar,method=PointSpreadFunctions.MethodRichardsWolf, aberrations= aberr, FFTPlan=FFTW.MEASURE)
PSFParams(0.58, 1.4, 1.518, Float32, ModeWidefield, PointSpreadFunctions.pol_scalar, PointSpreadFunctions.var"#42#43"(), PointSpreadFunctions.MethodSincR, 0x00000000, nothing, nothing)

PointSpreadFunctions.PSFParams — Type

PSFParams(λ=0.5, NA=1.2, n=1.33; pol=pol_circ, dtype=Float32, mode=ModeWidefield, 
aplanatic = aplanatic_detection, method=MethodRichardsWolf, FFTPlan=nothing,
aberrations=Aberrations(), pixelshape=nothing, transition_dipole=nothing, λ_weights=nothing)

creates the PSFParams structure via this constructor. You can also call PSFParams with the first argument being an existing structure and just specify the parameters to change via one or multiple named arguments.

Arguments:

λ: Vacuum emsision wavelength (same units as sampling, default is 0.5 µm)

NA: numerical aperture

n: refractive index of the embedding AND immersion medium

pol: a function calculating the polarization from a given Tuple of relative-k pupil vectors

dtype: real-valued data type to generate PSF for

mode: microscopy mode to calculate PSF for ::PSFMode.

ModeWidefield (default): Widefield microscopy

Mode2Photon: Two-photon microscopy

Mode4Pi: 4Pi microscopy

ModeConfocal: Confocal microscopy

ModeISM: Image scanning microscopy

ModeSTED: stimulated emission depletion microscopy

ModeLSCylinder: Light-sheet microscopy illuminating with a cylindrical lens

ModeDSLM: Dynamically scanned light-sheet microscopy

method: microscopy mode to calculate PSF for valid options are currently:

MethodPropagate: Angulare spectrum propagation. This version does NOT account for wrap around problems yielding problems at larger out-of-focus distances

MethodPropagateIterativ (default): Angulare spectrum propagation accounting from wrap-around problems in each propagation step by applying a perfectly matched layer (PML).

MethodShell: Angulare spectrum propagation with a slightly different calculation order. This version does NOT account for wrap around problems yielding problems at larger out-of-focus distances

MethodSincR: Based on first calculating a SincR function in real space and applying consecutive filtering steps. It accounts for wrap around problems but requires a quite high sampling along the Z direction.

MethodRichardsWolf: Uses the method described in the paper by B. Richards and E. Wolf, "Electromagnetic diffraction in optical systems. II. structure of the image field in an aplanatic system," Proc. R. Soc. London A, vol. 253, no. 1274, 1959. The terms I0, I1 and I2 are first calculated for an radial Z-dependet profile and then interpolated onto the 3D volume.

aplanatic: aplanatic factor. Provided as a function of angle θ. Choices are aplanatic_const, aplanatic_detection, aplanatic_illumination, aplanatic_illumination_flux

FFTPlan: information on how to calculate the FFTW plan. Default: nothing (using FFTW.ESTIMATE)

transition_dipole If supplied a transition-dipole (e.g. sqrt(2) .* (0.0,1.0,1.0)) will be accounted for in the PSF calculation. If not normalized, the strength will be included.

Example:

julia> using FFTW, PointSpreadFunctions

julia> ppm = PSFParams(0.58, 1.4, 1.518;pol=pol_circ,method=PointSpreadFunctions.MethodSincR, aberrations= Aberrations(), FFTPlan=FFTW.MEASURE)
PSFParams(0.58, 1.4, 1.518, Float32, ModeWidefield, PointSpreadFunctions.pol_scalar, PointSpreadFunctions.var"#42#43"(), PointSpreadFunctions.MethodSincR, 0x00000000, nothing, nothing)

julia> ppem = PSFParams(ppm; λ=0.620)

Defining Aberrations

PointSpreadFunctions.Aberrations — Type

Aberrations
defining Zernike phase aberrations via a list of indices and coefficients and an indexing style.